+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Deprecated `Pundit::SUFFIX`, moved it to `Pundit::PolicyFinder::SUFFIX` (#835)
 - Explicitly require less of `active_support` (#837)
+- Using `permit` matcher without a surrouding `permissions` block now raises a useful error. (#836)
 
 ## 2.4.0 (2024-08-26)